pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "clin-omics"
version = "0.1.0"
description = "Reusable framework for clinical and omics data integration, curation, and analysis."
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "Tadahaya Mizuno" },
]
license = { text = "MIT" }
dependencies = [
  "numpy",
  "pandas",
  "pyarrow",
  "h5py",
  "scikit-learn",
  "networkx",
  "matplotlib",
  "umap-learn",
]

[project.scripts]
clin-omics = "clin_omics.cli:main"

[project.optional-dependencies]
anndata = [
  "anndata",
]
leiden = [
  "igraph",
  "leidenalg",
]
dev = [
  "pytest",
  "black",
  "ruff",
  "isort",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"